+++ admin/users.php
@@ -149,6 +149,10 @@ switch ($_REQUEST['action']) {
 	break;
 	case 'confirm_delete':
 	        if (Config::get('demo_mode')) { break; }
+		if (!Core::form_verify('delete_user')) { 
+			access_denied(); 
+			exit; 
+		} 
 		$client = new User($_REQUEST['user_id']); 
 		if ($client->delete()) { 
 			show_confirmation(_('User Deleted'), "$client->username has been Deleted","admin/users.php");
@@ -162,7 +166,7 @@ switch ($_REQUEST['action']) {
 		$client = new User($_REQUEST['user_id']); 
 		show_confirmation(_('Deletion Request'),
 			_('Are you sure you want to permanently delete') . " $client->fullname ($client->username)?",
-			"admin/users.php?action=confirm_delete&amp;user_id=" . $_REQUEST['user_id'],1);
+			"admin/users.php?action=confirm_delete&amp;user_id=" . $client->id,1,'delete_user');
 	break;
 	/* Show IP History for the Specified User */
 	case 'show_ip_history':
